creative_color_model/requirements.txt
colorthief==0.2.1
numpy==1.24.3
Pillow==9.3.0
requests==2.28.1
